+++ tests/QueueManagerTest.php
@@ -243,7 +243,7 @@ class QueueManagerTest extends \PHPUnit_Framework_TestCase
         $this->assertFalse(array_key_exists('supplier_info', $data['positions'][0]['agent_info']));
         $this->assertEquals(
             json_encode($data['positions'][0]['nomenclature_code']),
-            '{"type":"shoes","code":"019876543210123421sgEKKPPcS25y5","byte_code":"444D00000096b43f303132333432317367454b4b5050635332357935"}'
+            '{"type":"shoes","code":"019876543210123421sgEKKPPcS25y5","hex_code":"444D00000096b43f303132333432317367454b4b5050635332357935"}'
         );
 
         $path = 'api/shop/v1/queues/queue-id/task';